Refactor: boolean 변수명 변경 및 코드 분리
- renderCodeTags함수와 renderContent함수로를 utils/readTimeTest.jsx로 분리하였습니다.
- 그에 따라 기존 readTimeTest.js 함수를 readTimeTest.jsx 함수로 변경하였습니다.
